backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
pydantic>=2.0,<3.0
mysql-connector-python>=9.0,<10.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
